backend/requirements.txt
flask
loguru
flask_cors
python-docx
openai
deepl